[IMP] tools,web,website: move `DotDict` and `MockRequest`
The `DotDict` class can produce unexpected behaviour
if used incorrectly.
This commit moves this class so that it is only used in tests.

The `MockRequest` context manager uses the `DotDict`.
As we cannot import objects from the test framework and
`MockRequest` is only used in tests, it makes sense to move it
to `odoo.addons.web.tests.utils`.

task-4822364
