+++ addons/web/tests/__init__.py
@@ -1,5 +1,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import utils
+
 from . import test_db_manager
 from . import test_health
 from . import test_image
