+++ addons/website_event/tests/test_event_internals.py
@@ -5,8 +5,8 @@ from datetime import datetime, timedelta
 
 from odoo.fields import Datetime as FieldsDatetime
 from odoo.tests.common import users
+from odoo.addons.web.tests.utils import MockRequest
 from odoo.addons.website.tests.test_website_visitor import MockVisitor
-from odoo.addons.website.tools import MockRequest
 from odoo.addons.website_event.controllers.main import WebsiteEventController
 from odoo.addons.event.tests.common import EventCase
