+++ addons/website_sale/tests/test_website_editor.py
@@ -7,7 +7,7 @@ from odoo.exceptions import ValidationError
 from odoo.fields import Command
 from odoo.tests import HttpCase, tagged
 
-from odoo.addons.website.tools import MockRequest
+from odoo.addons.web.tests.utils import MockRequest
 from odoo.addons.website_sale.controllers.main import WebsiteSale
